Documents/Projects/FashionFit/FashionFitLocal/project/backend/requirements.txt
flask
paypalrestsdk
python-dotenv
requests
pymongo
flask-cors